+++ bcipy/helpers/stimuli.py
