+++ bcipy/helpers/tests/test_signal_model.py
@@ -1,8 +1,8 @@
 import unittest
 import numpy as np
 
-from bcipy.helpers.signal_model_related import CopyPhraseWrapper
-from bcipy.helpers.bci_task_related import alphabet
+from bcipy.helpers.signal_model import CopyPhraseWrapper
+from bcipy.helpers.task import alphabet
 
 
 class TestSignalModelRelated(unittest.TestCase):
