+++ bcipy/language_model/demo/prelm_demo.py
@@ -1,7 +1,7 @@
 import sys
 sys.path.append('.')
 from bcipy.language_model.prelm_language_model import LangModel
-from bcipy.helpers.lang_model_related import norm_domain
+from bcipy.helpers.lang_model import norm_domain
 
 
 def main():
